Update 6x50XT cards
Add: 2 new 6650XT cards
 - https://www.microcenter.com/product/649013/asus-amd-radeon-rx-6650-xt-rog-strix-overclocked-dual-fan-8gb-gddr6-pcie-40-graphics-card (6650 - SKU 397190)
 - https://www.microcenter.com/product/649014/asus-amd-radeon-rx-6650-xt-dual-overclocked-dual-fan-8gb-gddr6-pcie-40-graphics-card (6650 - SKU 397083)

Also changed the names of some other cards (like the Liquid Devil and Red Devils) in order to make them fit in with the list better. For example, changed RED DEVIL to Red Devil - minor but nice changes.

Additional note:
As for the Gigabyte 3080TI gaming, the SKU is the same - just some weird issues with the bot.